Avoid use of long double in cubic
Cubic uses long double which is not standardized. On x86-64 it provides
only 80 bits of precision, but on other architectures it provides either
64 bits or 128 bits of precision.

This not only means that we are doing different computations, but we
are also unfairly penalising architectures that use a 128 bit long
double. Those architectures often have a software emulation library for
128 bit floating point arithmetic. It was also pointed out to me that
the benchmark isn't consistently using long double - it's using double
versions of the transcendental functions.

Fix this by converting all uses of long double to double. The benchmark
still verifies with this change.

ChangeLog:

        * src/cubic/libcubic.c (SolveCubic): Replace long double with
        double.