Fix benchmark_size.py script to select duplicate and derivative names of sections for the size calculation.
When compiled with options like -ffunction-sections or similar or
in the presence of particular linker scripts, some sections might
have similar names (e.g.: .text, .text.benchmark_body, etc... ).
The script was previously designed to take only the main sections
into consideration (.text, .rodata, .data, .bss) as the default
linker scripts of GCC for ARM and RISC-V group all the relevant
sections into such main ones, regardless of the usage of
-ffunction-sections or -fdata-sections or similar.
However that might not be the case for other compilers or linker
scripts.

The script now selects the section mentioned with the 'metric'
parameter (or .text as a default) and all the subsections of the
same.

Files changed:

        * benchmark_size.py: Added a search for the affiliate
          sections. Removed the encoding of the name of the
          section as it caused error.