Add files to enable Apple Intel Mac to run embench-iot speed benchmarks.
Size benchmarks are trickier as the Mac uses a different, non-ELF file format.

These changes support both Intel x86 and Apple ARM macs with the capability of
building the benchmarks for x86 and Apple on either sort of Mac host, running
benchmarks natively, and of runnning benchmarks compiled for x86 on Apple
Silicon (via Rosetta).

The way that architecture independence is implemented means that is not
straightforward to use the Embench scripts to benchmark natively. The problem
arises when the Python system is an x86 executable running on an Apple Silicon
Mac. When a process is spawned - for example to compile some code - the
environment that process sees is x86, and the development tools will assume,
unless directed otherwise, that they are building code for an x86 Mac.

Accordingly there are three flavours of build (for speed) files under config/mac:
        chips/speed-test-clang which compiles for the "native" architecture
        chips/speed-test-clang-ARM-M1 which compiles for Apple Silicon
        chips/speed-test-cland-X86-64 which compiles for x86

The speed benchmarks use the BSD timer to measure execution time. The reported
time is real elasped time, the benchmark also measures "CPU time" but this is
not used to produce the results.

Files created:

        * config/mac/README.md
        * config/mac/arch.cfg
        * config/mac/boards/generic/board.cfg
        * config/mac/boards/generic/boardsupport.c
        * config/mac/boards/generic/boardsupport.h
        * config/mac/chips/generic/chip.cfg
        * config/mac/chips/generic/chipsupport.c
        * config/mac/chips/generic/chipsupport.h
        * config/mac/chips/size-test-clang/chip.cfg
        * config/mac/chips/size-test-clang/chipsupport.c
        * config/mac/chips/size-test-clang/chipsupport.h
        * config/mac/chips/speed-test-clang-ARM-M1/chip.cfg
        * config/mac/chips/speed-test-clang-ARM-M1/chipsupport.c
        * config/mac/chips/speed-test-clang-ARM-M1/chipsupport.h
        * config/mac/chips/speed-test-clang-X86-64/chip.cfg
        * config/mac/chips/speed-test-clang-X86-64/chipsupport.c
        * config/mac/chips/speed-test-clang-X86-64/chipsupport.h
        * config/mac/chips/speed-test-clang/chip.cfg
        * config/mac/chips/speed-test-clang/chipsupport.c
        * config/mac/chips/speed-test-clang/chipsupport.h
        * pylib/run_mac.py