Fix to issue #113 pt 1 (#115)
* This edit deals with issue #113 pt 1 as sugegsted in the inital comment.

* This concerns issue #113 pt 1 - replacing start_benchmark and stop_benchmark
with start_trigger and stop_trigger.

The previous commit which was pushed to my forked git-hub repository was
incorrect and deleted a lot of necessary text from the file. I think this
was caused by the toc insertion not working correctly.

Accordlingly this commit is used the previous version of this file
and only performed the replacement. NEITHER spelling checking NOR
toc generation have been performed.