Improve JSON output of benchmarking
Currently when benchmarking with `--json-output` invalid `json` is
produced. This is because surrounding braces are missing AND some debug
output is produced saying that the benchmarks were run successfully.

When users want to get JSON output, it's supposed to be machine-readable
as opposed to human-readable output. Therefore, this commits fixes the
JSON output, so that the benchmarks output valid JSON and don't dump
debug output.

This also splits out the printing logic from the collecting the data
logic. Collecting data and outputting them should be separated IMO.