Add primecount test.
The test primecount searches and counts the prime numbers up to a given limit.
This version of the test is a reduction of the test proposed by Bruce Hoult.
The limit has been reduced in order to fit the execution time on a Cortex M4
working at 16 MHz into 4 seconds.

Files changed:

	* src/primecount/primecount.c: Created.
	* baseline-data/size.json: Added the result for primecount.
	* baseline-data/speed.json: Likewise.

Thanks to Bruce Hoult <bruce@hoult.org> for providing the test.