[MD5SUM] Fix local scale factor, heap size, add iterations and add baseline values.
The scale factor and heap size needed to be adjusted in order to fit on the
baseline board (embench-0.5).
The previous choice to scale the workload of the test through the local scale
factor by increasing the size of the message to be processed imposed a limit
on the scalability of the test as above a certain memory requirement the test
went from taking tens of milliseconds to hanging indeterminately.
This solution fixes a determined length of 1000 bytes for the message and
repeats the operation (LOCAL_SCALE_FACTOR*CPU_MHZ) times, as it usually
happens for the other benchmarks.
As a consequence of this also the heap size requirement is smaller and such
size was set arbitrarily to 2000, twice the size of the message, just as a
buffer for architectural differences of different targets.

Files changed:

	* baseline-data/speed.json: Added baseline speed for md5sum.
	* baseline-data/size.json: Added baseline size for md5sum.
	* src/md5sum/md5.c: Fixed local scale factor, added iterations,
	fixed heap size.