[TARFIND] Fix local scale factor, heap size, add iterations and add baseline values.
The number of iterations was limited by the number of files in the archive as
this determined the required amount of heap.
Now iterations have been added on the whole benchmark body and such iterations
are determined freely by the local scale factor and the cpu frequency.
The initialization of the heap and of the files in the archive has been moved
in the part of benchmark that is iterated, like the other benchmarks.
The number of files in the archive and consecutively the requirement of heap
size has been reduced so that it doesn't exceed the average heap requirement
of the other tests and the initialization part has smaller influence then the
search of the files in the archive, that is the core of the benchmark.

Files changed:

	* baseline-data/speed.json: Added baseline speed for tarfind.
	* baseline-data/size.json: Added baseline size for tarfind.
	* src/tarfind/tarfind.c: Fixed local scale factor, added iterations,
	fixed heap size.